hotels

+++ Despegar.Core.Business/Hotels/BookingFields/Installments.cs
@@ -7,7 +7,7 @@ namespace Despegar.Core.Business.Hotels.BookingFields
 {
     public class Installments
     {
-        public decimal quantity { get; set; }
+        public int quantity { get; set; }
         public decimal first { get; set; }
         public decimal others { get; set; }
     }
